Update rest service to signal camera thread runner.
The rest service is the entry point of this application, hence the hook to
trigger the camera thread service is added into the rest service module.

After each successful operation on the system, the rest service now signal
the threadrunner as well to operate on threads accordingly.
This signal make create, delete or update the parameters in the existing
threads.

Signed-off-by: sugu <sugeshchandran@gmail.com>